# Implement optimized batch processing for apply_predictions_to_dataframe() - Issue #10